# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Top Music API"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": ["http"],
  "paths": {
...
    "/songs": {
      "get": {
        "tags": ["songs"],
        "summary": "Get songs",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "description": "Filter songs by genre",
            "type": "string"
          },
          {
            "name": "danceability",
            "in": "query",
            "description": "Filter songs by danceability",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of songs",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "body": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Song"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/songs/artist/{artistName}": {
      "get": {
        "tags": ["songs"],
        "summary": "Get songs by artist",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "artistName",
            "in": "path",
            "description": "Name of the artist",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of songs",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "body": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Song"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Song": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "artist": {
          "type": "string"
        },
        "genre": {
          "type": "string"
        },
        "danceability": {
          "type": "number"
        }
      }
    }
  }
}